Convert old block mapping JSON to new one
