+++ src/main/java/org/geysermc/generator/Main.java
@@ -1,7 +1,5 @@
 package org.geysermc.generator;
 
-import org.geysermc.generator.state.CollisionGenerator;
-
 public class Main {
 
     public static void main(String[] args) {
@@ -9,13 +7,13 @@ public class Main {
 
         MappingsGenerator generator = new MappingsGenerator();
         generator.generateItems();
-        generator.generateBlocks();
+        BlockGenerator.generate();
         CollisionGenerator.generate();
         generator.generateSounds();
         generator.generateBiomes();
         generator.generateMapColors();
         generator.generateEnchantments();
         generator.generateParticles();
-        generator.generateInteractionData();
+        //generator.generateInteractionData();
     }
 }
